requirements.txt
faiss_cpu==1.10.0
google_api_python_client==2.166.0
huggingface_hub==0.29.3
numpy==2.2.4
pandas==2.2.3
protobuf==6.30.2
streamlit==1.44.1
together==1.5.5
torch==2.6.0
transformers==4.50.0
